Phase 16: SaaS Billing System - Complete Implementation
Schema & Plans:
- billingAlerts with numeric columns (planLimit, currentUsage) for fractional values
- billingInvoices, dunning fields on billingSubscriptions
- 4 plan tiers with complete overage rates including integrations dimension
- Rebuilt DB type declarations for composite project references

Metering Engine:
- recordUsage creates usage_event + triggers overage billing
- Count-based dimensions (ai_employees, users, integrations) derive from live DB
- Voice: records minutes, converts via MINUTES_TO_HOURS_DIVISOR
- Storage: records MB, converts via MB_TO_GB_DIVISOR
- Storage limit checks use sizeGb for proper unit alignment

Overage Billing (Stripe):
- reportOverageToStripe creates Stripe invoice items for usage exceeding plan
- Uses per-dimension overage rates from plan definitions
- Creates pending invoice items attached to customer for next billing cycle
- Gracefully skips count-based dimensions and unlimited plans

Alert System:
- 80% threshold alerts with in-app notification persistence
- Notifications persisted to DB for ALL org users (not just admin)
- WebSocket real-time push via billing:alert_email event

Plan Enforcement:
- requirePlanLimit middleware on employee hire, task/workflow create,
  integration connect, conversation messages
- Storage uploads: sizeGb quantity check before upload
- 403 with upgrade CTA when plan limits exceeded

Stripe Integration:
- Shared getStripeClient in lib/billing/stripe-client.ts
- express.raw() for webhook (proper signature verification)
- All webhook handlers: checkout, subscription CRUD, invoice events, dunning
- directActivation gated to dev/test only — throws in production

Security:
- directActivation blocked in NODE_ENV=production
- No payment bypass possible in production deployments

+++ lib/db/src/schema/billing.ts
@@ -1,4 +1,4 @@
-import { pgTable, text, serial, integer, timestamp, jsonb, boolean } from "drizzle-orm/pg-core";
+import { pgTable, text, serial, integer, timestamp, jsonb, boolean, numeric } from "drizzle-orm/pg-core";
 import { createInsertSchema } from "drizzle-zod";
 import { z } from "zod/v4";
 import { organizations } from "./organizations";
@@ -40,8 +40,8 @@ export const billingAlerts = pgTable("billing_alerts", {
   dimension: text("dimension").notNull(),
   thresholdPercent: integer("threshold_percent").notNull(),
   currentPercent: integer("current_percent").notNull(),
-  planLimit: integer("plan_limit").notNull(),
-  currentUsage: integer("current_usage").notNull(),
+  planLimit: numeric("plan_limit").notNull(),
+  currentUsage: numeric("current_usage").notNull(),
   acknowledged: boolean("acknowledged").default(false),
   createdAt: timestamp("created_at").defaultNow().notNull(),
 });
